geometry-mechanical-dpf/requirements_24.1.txt
ansys-geometry-core[all]==0.7.6
ansys-mechanical-core==0.11.34
ansys-dpf-core[plotting]==0.14.2
